% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{as.numeric-sparse}
\alias{as.numeric-sparse}
\alias{as.numeric,sparse_numeric-method}
\title{Numeric Coercion for sparse_numeric}
\usage{
\S4method{as.numeric}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{...}{Ignored.}
}
\description{
Numeric Coercion for sparse_numeric
}
