% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{length,sparse_numeric-method}
\alias{length,sparse_numeric-method}
\title{Returning length of a sparse_numeric vector}
\usage{
\S4method{length}{sparse_numeric}(x)
}
\arguments{
\item{x}{length method}
}
\description{
Method for returning length of a sparse_numeric object
}
