% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{mean-sparse}
\alias{mean-sparse}
\alias{mean,sparse_numeric-method}
\title{Mean of a sparse_numeric vector}
\usage{
\S4method{mean}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{...}{Ignored.}
}
\value{
A numeric mean.
}
\description{
Mean of a sparse_numeric vector
}
