% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{mean,sparse_numeric-method}
\alias{mean,sparse_numeric-method}
\title{Finding the mean of a sparse_numeric vector}
\usage{
\S4method{mean}{sparse_numeric}(x)
}
\arguments{
\item{x}{mean method}
}
\description{
Method for finding the mean of a sparse_numeric object
}
