% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mode}
\alias{mode}
\title{mode}
\usage{
mode(v)
}
\arguments{
\item{v}{- integer vector}
}
\value{
integer value
}
\description{
This function determine the most frequently occurring value in an integer vector (mode). If the mode is ambiguous (e.g. for 1, 2, 2, 2, 3, 3, 1, 3), return any mode
}

