\name{Homework 6}
\alias{mypackage}
\docType{package}
\title{
  A package containing Homework 6
}
\description{
The package contains homework's functions.
}
\details{
  The package contains a function 'mode' which determines the most frequently occurring value in an numeric vector (mode). If the mode is ambiguous (e.g. for 1, 2, 2, 2, 3, 3, 1, 3), returns any mode.
}
\author{
Michal Pawlowski, michciek@gmail.com
}

\examples{
  mode(c(1,2,3,pi,pi))
}
