% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{norm-sparse}
\alias{norm-sparse}
\alias{norm,sparse_numeric,ANY-method}
\title{Euclidean Norm of a sparse_numeric vector}
\usage{
\S4method{norm}{sparse_numeric,ANY}(x, type = "2", ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{type}{Ignored (kept for compatibility).}

\item{...}{Ignored.}
}
\value{
Numeric norm.
}
\description{
Euclidean Norm of a sparse_numeric vector
}
