% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{norm,sparse_numeric-method}
\alias{norm,sparse_numeric-method}
\title{Finding squared norm of a sparse_numeric vector}
\usage{
\S4method{norm}{sparse_numeric}(x)
}
\arguments{
\item{x}{norm method}
}
\description{
Method for finding squared norm of a sparse_numeric object
}
