% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{plot-sparse}
\alias{plot-sparse}
\alias{plot,sparse_numeric,sparse_numeric-method}
\title{Plot Two Sparse Vectors}
\usage{
\S4method{plot}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{y}{A \code{sparse_numeric} object.}

\item{...}{Passed to \code{plot()}.}
}
\description{
Plot Two Sparse Vectors
}
