% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{plot,sparse_numeric,sparse_numeric-method}
\alias{plot,sparse_numeric,sparse_numeric-method}
\title{Plotting two sparse_numeric vectors}
\usage{
\S4method{plot}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{plot method}

\item{y}{plot method}
}
\description{
Method for plotting two sparse_numeric vectors
}
