% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework6.R
\name{pythagorean}
\alias{pythagorean}
\title{pythagorean}
\usage{
pythagorean(a, b, c)
}
\arguments{
\item{a}{a side leg of a right triangle}

\item{b}{a side leg of a right triangle}

\item{c}{hypotenuse of a right triangle}
}
\value{
The length of the rest side or hypotenuse
}
\description{
It is used to calculate a length of side leg a, b or hypotenuse for right triagles.
}
\examples{
pythagorean(a=3,b=4)
}
