% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{show-sparse}
\alias{show-sparse}
\alias{show,sparse_numeric-method}
\title{Show Method for sparse_numeric}
\usage{
\S4method{show}{sparse_numeric}(object)
}
\arguments{
\item{object}{A \code{sparse_numeric} object.}
}
\description{
Show Method for sparse_numeric
}
