% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{show,sparse_numeric-method}
\alias{show,sparse_numeric-method}
\title{Printing a sparse_numeric vector}
\usage{
\S4method{show}{sparse_numeric}(object)
}
\arguments{
\item{object}{show method}
}
\description{
Method for printing a sparse_numeric object
}
