% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simplify2array}
\alias{simplify2array}
\title{simplify2array}
\usage{
simplify2array(x)
}
\arguments{
\item{x}{- list of numeric vectors}
}
\value{
matrix
}
\description{
This function filled  matrix by vectors column-by-column.
}

