% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{sparse_add,sparse_numeric,sparse_numeric-method}
\alias{sparse_add,sparse_numeric,sparse_numeric-method}
\title{Adding two sparse_numeric vectors}
\usage{
\S4method{sparse_add}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{method sparse_add}

\item{y}{method sparse_add}
}
\description{
Method for adding two sparse_numeric vectors
}
