% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{sparse_add}
\alias{sparse_add}
\title{Add two sparse_numeric vectors}
\usage{
sparse_add(x, y)
}
\arguments{
\item{x}{generic sparse_add}

\item{y}{generic sparse_add}
}
\description{
Generic function for adding two sparse_numeric objects
}
