% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_crossprod-method}
\alias{sparse_crossprod-method}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\title{Dot Product of Two Sparse Vectors}
\usage{
\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{y}{A \code{sparse_numeric} object.}

\item{...}{Additional arguments (unused).}
}
\value{
A single numeric value.
}
\description{
Dot Product of Two Sparse Vectors
}
