% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\title{Finding cross-product of two sparse_numeric vectors}
\usage{
\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{method sparse_crossprod}

\item{y}{method sparse_crossprod}
}
\description{
Method for finding cross-product of two sparse_numeric vectors
}
