% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\title{Finding cross-product of two sparse_numeric vectors}
\usage{
sparse_crossprod(x, y)
}
\arguments{
\item{x}{generic sparse_crossprod}

\item{y}{generic sparse_crossprod}
}
\description{
Generic function for finding cross-product of two sparse_numeric objects
}
