% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{sparse_mult,sparse_numeric,sparse_numeric-method}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\title{Multiplying two sparse_numeric vectors}
\usage{
\S4method{sparse_mult}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{method sparse_mult}

\item{y}{method sparse_mult}
}
\description{
Method for multiplying two sparse_numeric vectors
}
