% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{sparse_mult}
\alias{sparse_mult}
\title{Multiplying two sparse_numeric vectors}
\usage{
sparse_mult(x, y)
}
\arguments{
\item{x}{generic sparse_mult}

\item{y}{generic sparse_mult}
}
\description{
Generic function for multiplying two sparse_numeric objects
}
