% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_numeric-coerce}
\alias{sparse_numeric-coerce}
\alias{coerce,numeric,sparse_numeric-method}
\alias{coerce,sparse_numeric,numeric-method}
\title{Coercion to/from sparse_numeric}
\arguments{
\item{from}{Object to convert.}
}
\value{
A \code{sparse_numeric} object or numeric vector.
}
\description{
Coercion to/from sparse_numeric
}
