% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric-methods-doc.R,
%   R/sparse_numeric.R
\docType{methods}
\name{sparse_numeric-methods}
\alias{sparse_numeric-methods}
\alias{show,sparse_numeric-method}
\alias{sparse_add,sparse_numeric,sparse_numeric-method}
\alias{sparse_sub,sparse_numeric,sparse_numeric-method}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\alias{+,sparse_numeric,sparse_numeric-method}
\alias{-,sparse_numeric,sparse_numeric-method}
\alias{*,sparse_numeric,sparse_numeric-method}
\alias{norm,sparse_numeric-method}
\alias{standardize,sparse_numeric-method}
\alias{plot,sparse_numeric,sparse_numeric-method}
\alias{length,sparse_numeric-method}
\title{Methods for sparse_numeric objects}
\usage{
\S4method{show}{sparse_numeric}(object)

\S4method{sparse_add}{sparse_numeric,sparse_numeric}(x, y)

\S4method{sparse_sub}{sparse_numeric,sparse_numeric}(x, y)

\S4method{sparse_mult}{sparse_numeric,sparse_numeric}(x, y)

\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(x, y)

\S4method{+}{sparse_numeric,sparse_numeric}(e1, e2)

\S4method{-}{sparse_numeric,sparse_numeric}(e1, e2)

\S4method{*}{sparse_numeric,sparse_numeric}(e1, e2)

\S4method{norm}{sparse_numeric}(x, ...)

\S4method{standardize}{sparse_numeric}(x, ...)

\S4method{plot}{sparse_numeric,sparse_numeric}(x, y, ...)

\S4method{length}{sparse_numeric}(x)
}
\arguments{
\item{object}{A \code{sparse_numeric} object.}

\item{x, y, e1, e2}{\code{sparse_numeric} objects used as operands in
arithmetic or utility methods.}

\item{...}{Additional arguments passed to methods (currently ignored).}
}
\description{
S4 methods defined for the \code{sparse_numeric} class, including
arithmetic operators, norms, and utility functions.
}
\details{
This help topic exists primarily to satisfy R CMD check that these
methods are documented. See the documentation for the
\code{sparse_numeric} class and the corresponding generics for
details on behavior.
}
