% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{-,sparse_numeric,sparse_numeric-method}
\alias{-,sparse_numeric,sparse_numeric-method}
\title{Subtracting two sparse_numeric vectors}
\usage{
\S4method{-}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{e1}{arithmetic "-"}

\item{e2}{arithmetic "-"}
}
\description{
Method for subtracting two sparse_numeric vectors with "-"
}
