% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_numeric_add_op}
\alias{sparse_numeric_add_op}
\alias{+,sparse_numeric,sparse_numeric-method}
\title{Addition operator for sparse_numeric}
\usage{
\S4method{+}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{e1}{A \code{sparse_numeric} object.}

\item{e2}{A \code{sparse_numeric} object.}
}
\value{
A new \code{sparse_numeric} vector.
}
\description{
Addition operator for sparse_numeric
}
