% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{sparse_sub-method}
\alias{sparse_sub-method}
\alias{sparse_sub,sparse_numeric,sparse_numeric-method}
\title{Subtract Two Sparse Numeric Vectors}
\usage{
\S4method{sparse_sub}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{y}{A \code{sparse_numeric} object.}

\item{...}{Additional arguments (unused).}
}
\value{
A new \code{sparse_numeric} vector.
}
\description{
Subtract Two Sparse Numeric Vectors
}
