% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{sparse_sub,sparse_numeric,sparse_numeric-method}
\alias{sparse_sub,sparse_numeric,sparse_numeric-method}
\title{Subtracting two sparse_numeric vectors}
\usage{
\S4method{sparse_sub}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{method sparse_sub}

\item{y}{method sparse_sub}
}
\description{
Method for subtracting two sparse_numeric vectors
}
