% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{sparse_sub}
\alias{sparse_sub}
\title{Subtracting two sparse_numeric vectors}
\usage{
sparse_sub(x, y)
}
\arguments{
\item{x}{generic sparse_sub}

\item{y}{generic sparse_sub}
}
\description{
Generic function for subtracting two sparse_numeric objects
}
