% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW5_Class.R
\name{standardize-method}
\alias{standardize-method}
\alias{standardize,sparse_numeric-method}
\title{Standardize a Sparse Numeric Vector}
\usage{
\S4method{standardize}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A sparse_numeric vector.}

\item{...}{Additional arguments (unused).}
}
\value{
A standardized sparse_numeric object.
}
\description{
Standardize a Sparse Numeric Vector
}
