% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{standardize,sparse_numeric-method}
\alias{standardize,sparse_numeric-method}
\title{Standardizing a sparse_numeric vector}
\usage{
\S4method{standardize}{sparse_numeric}(x)
}
\arguments{
\item{x}{standardize method}
}
\description{
Method for standardizing a sparse_numeric object
}
