% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{standardize}
\alias{standardize}
\title{Standardize a sparse_numeric vector}
\usage{
standardize(x, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} vector.}

\item{...}{Ignored.}
}
\value{
A new standardized \code{sparse_numeric} vector.
}
\description{
Standardizes a \code{sparse_numeric} vector by subtracting the mean
and dividing by the standard deviation. Implicit zeros are included
in the calculations by converting to a dense numeric vector and
then back to \code{sparse_numeric}.
}
