% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sum,sparse_numeric-method}
\alias{sum,sparse_numeric-method}
\title{Sum of Sparse Numeric Vector}
\usage{
\S4method{sum}{sparse_numeric}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{sparse_numeric object}

\item{...}{Additional arguments passed to sum}

\item{na.rm}{logical, should NA values be removed}
}
\description{
Sum of Sparse Numeric Vector
}
