% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework6.R
\name{trimmed_mean}
\alias{trimmed_mean}
\title{trimmed_mean}
\usage{
trimmed_mean(x, s, l)
}
\arguments{
\item{x}{A vector of numeric values}

\item{s}{The s smallest values to cut off}

\item{l}{The l largest values to cut off}
}
\value{
The average of a trimmed vector x
}
\description{
It is used to find a trimmed mean.
}
\examples{
x<-c(1,7,3,2,5,0.5,9,10)
trimmed_mean(x,s=1,l=2)
## trimmed vector with s=1,l=2 is c(1,2,3,5,7)
}
