% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{package}
\name{Homework6-package}
\alias{Homework6}
\alias{Homework6-package}
\title{Homework6: Sparse Numeric Vector Tools}
\description{
Provides an S4 class for sparse numeric vectors and basic operations
such as arithmetic, norms, and standardization without converting
to dense format.
}
\author{
\strong{Maintainer}: Jana Kassem \email{jana.kassem@utexas.edu}

}
\keyword{internal}
