% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework6.R
\name{mean,sparse_numeric-method}
\alias{mean,sparse_numeric-method}
\title{Mean of a sparse_numeric vector}
\usage{
\S4method{mean}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric}}

\item{...}{Ignored}
}
\value{
A numeric scalar giving mean of vector.
}
\description{
Computes the mean of a \code{sparse_numeric} vector, counting the stored
non-zeros
}
