% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{mean}
\alias{mean}
\alias{mean,sparse_numeric-method}
\title{Mean of Sparse Numeric Vector}
\usage{
mean(x, ...)

\S4method{mean}{sparse_numeric}(x, trim = 0, na.rm = FALSE, ...)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{...}{Additional arguments (not used)}

\item{trim}{trim value}

\item{na.rm}{logical, should missing values be removed?}
}
\value{
mean of vector
}
\description{
computes mean of sparse numeric vector
}
