% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{norm}
\alias{norm}
\alias{norm,sparse_numeric-method}
\title{Norm of Sparse Numeric Vector}
\usage{
norm(x, ...)

\S4method{norm}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{...}{Additional arguments (not used)}
}
\value{
The  norm of the vector
}
\description{
Computes norm of sparse vector
}
