% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{norm}
\alias{norm}
\title{Find squared norm of a sparse_numeric vector}
\usage{
norm(x)
}
\arguments{
\item{x}{generic sparse_norm}
}
\description{
Generic function for finding squared norm of a sparse_numeric vector
}
