% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework6.R
\name{norm}
\alias{norm}
\alias{norm,sparse_numeric-method}
\title{Euclidean norm of sparse_numeric vector}
\usage{
norm(x, ...)

\S4method{norm}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{...}{Ignored.}
}
\value{
A numeric scalar giving Euclidean norm
}
\description{
Computes norm of a \code{sparse_numeric} object, defined as square root of
the sum of squared elements
}
