% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6.R
\name{norm}
\alias{norm}
\alias{norm,sparse_numeric-method}
\title{computes the norm of a sparse vector}
\usage{
norm(x, ...)

\S4method{norm}{sparse_numeric}(x)
}
\arguments{
\item{x}{The sparse vector}

\item{...}{Not Necessary}
}
\value{
numeric number
}
\description{
This function normalizes a sparse vector, by computing the squared norm of the vector
}
