% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{norm}
\alias{norm}
\title{Euclidean norm of a sparse_numeric vector}
\usage{
norm(x, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{...}{Ignored.}
}
\value{
Numeric scalar norm.
}
\description{
Computes the Euclidean (L2) norm of a \code{sparse_numeric} vector.
}
