% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework6.R
\name{plot,sparse_numeric,sparse_numeric-method}
\alias{plot,sparse_numeric,sparse_numeric-method}
\title{Plot overlapping non-zero entries of two sparse_numeric vectors}
\usage{
\S4method{plot}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object.}

\item{y}{A \code{sparse_numeric} object.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A \code{ggplot} object is printed to the active graphics device.
}
\description{
Creates a scatter plot of the overlapping non-zero entries of two
\code{sparse_numeric} vectors of the same length.
}
