% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6.R
\name{plot,sparse_numeric,sparse_numeric-method}
\alias{plot,sparse_numeric,sparse_numeric-method}
\title{plots two sparse vectors}
\usage{
\S4method{plot}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{The first sparse vector}

\item{y}{The second sparse vector, of the same length as the first}
}
\value{
plot
}
\description{
This function plots the overlapping non-zero elements of two sparse vectors, returning the plot
}
