% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework6.R
\name{show,sparse_numeric-method}
\alias{show,sparse_numeric-method}
\title{Show a sparse_numeric object}
\usage{
\S4method{show}{sparse_numeric}(object)
}
\arguments{
\item{object}{a \code{sparse_numeric} vector}
}
\value{
Invisibly returns a \code{sparse_numeric} vector
}
\description{
Displays the content of a \code{sparse_numeric} vector, printing both
non-zero and zero elements as well as their position.
}
