% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6.R
\name{show,sparse_numeric-method}
\alias{show,sparse_numeric-method}
\title{shows a sparse vector}
\usage{
\S4method{show}{sparse_numeric}(object)
}
\arguments{
\item{object}{The first sparse vector}
}
\value{
dataframe of the sparse vector
}
\description{
This function outputs a sparse vector as a dataframe
}
