% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6.R
\name{sparse_add}
\alias{sparse_add}
\alias{sparse_add,sparse_numeric,sparse_numeric-method}
\title{Add two sparse vectors together}
\usage{
sparse_add(first, second, ...)

\S4method{sparse_add}{sparse_numeric,sparse_numeric}(first, second)
}
\arguments{
\item{first}{The first sparse vector}

\item{second}{The second sparse vector, of the same length as the first}

\item{...}{Not Necessary}
}
\value{
Sparse vector of the Sum
}
\description{
This function combines two sparse vectors of the same length, and then returns a sparse vector of the sum
}
