% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_add}
\alias{sparse_add}
\title{Add two sparse_numeric vectors}
\usage{
sparse_add(x, y)
}
\arguments{
\item{x, y}{Objects of class \code{sparse_numeric}.}
}
\value{
A \code{sparse_numeric} object of the same length.
}
\description{
Adds two \code{sparse_numeric} vectors of the same length and
returns their sum, represented as a \code{sparse_numeric}.
}
