% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\title{Dot Product/Cross Product (?) of Sparse Numeric Vectors}
\usage{
sparse_crossprod(x, y, ...)

\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{sparse_numeric object}

\item{y}{sparse_numeric object}

\item{...}{Additional arguments, not used}
}
\value{
value representing dot product
}
\description{
dot product of two sparse numeric vectors.
}
