% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework6.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\title{Cross multiplies two sparse numeric vectors together}
\usage{
sparse_crossprod(x, y, ...)

\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{a \code{sparse_numeric} vector}

\item{y}{a \code{sparse_numeric} vector}

\item{...}{ignored.}
}
\value{
A numeric scalar representing cross product of x and y
}
\description{
Computes the sum of all pairwise products between the non-zero entries of
two \code{sparse_numeric} vectors (a simple cross-product).
}
