% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\title{Inner product of two sparse_numeric vectors}
\usage{
sparse_crossprod(x, y)
}
\arguments{
\item{x, y}{Objects of class \code{sparse_numeric}.}
}
\value{
A numeric scalar giving the inner product.
}
\description{
Computes the standard Euclidean inner product between two
\code{sparse_numeric} vectors of the same length.
}
