% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework6.R
\name{sparse_mult}
\alias{sparse_mult}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\alias{*,sparse_numeric,sparse_numeric-method}
\title{Multiplies two sparse numeric vectors together}
\usage{
sparse_mult(x, y, ...)

\S4method{sparse_mult}{sparse_numeric,sparse_numeric}(x, y)

\S4method{*}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{x}{a \code{sparse_numeric} vector}

\item{y}{a \code{sparse_numeric} vector}

\item{...}{ignored.}

\item{e1}{A \code{sparse_numeric} object (left-hand operand for \code{*}).}

\item{e2}{A \code{sparse_numeric} object (right-hand operand for \code{*}).}
}
\value{
A \code{sparse_numeric} vector representing x * y (element-wise)
}
\description{
Multiplies two \code{sparse_numeric} vectors of the same length element-wise,
returning a new \code{sparse_numeric} object containing the non-zero elements
of the product
}
