% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6.R
\name{sparse_mult}
\alias{sparse_mult}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\title{Multiply two sparse vectors together}
\usage{
sparse_mult(first, second, ...)

\S4method{sparse_mult}{sparse_numeric,sparse_numeric}(first, second)
}
\arguments{
\item{first}{The first sparse vector}

\item{second}{The second sparse vector, of the same length as the first}

\item{...}{Not Necessary}
}
\value{
Sparse vector of the product of the first vector x the second vector
}
\description{
This function combines two sparse vectors of the same length, and then returns a sparse vector of their product
}
