% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_mult}
\alias{sparse_mult}
\title{Elementwise multiply two sparse_numeric vectors}
\usage{
sparse_mult(x, y)
}
\arguments{
\item{x, y}{Objects of class \code{sparse_numeric}.}
}
\value{
A \code{sparse_numeric} object of the same length.
}
\description{
Computes the elementwise (Hadamard) product \code{x * y} for two
\code{sparse_numeric} vectors of the same length.
}
