% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\docType{class}
\name{sparse_numeric-class}
\alias{sparse_numeric-class}
\title{Sparse Numeric Vector Class}
\description{
S4 class that represents vectors as sparse vectors, storing only nonzero values and positions.
}
\section{Slots}{

\describe{
\item{\code{value}}{nonzero values vector}

\item{\code{pos}}{nonzero value positions vector}

\item{\code{length}}{integer representing length of vector}
}}

