% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\docType{class}
\name{sparse_numeric-class}
\alias{sparse_numeric-class}
\title{sparse_numeric Class}
\description{
sparse_numeric Class
}
\section{Slots}{

\describe{
\item{\code{value}}{non-zero values}

\item{\code{pos}}{positions of values}

\item{\code{length}}{length of vector}
}}

