% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework6.R
\docType{class}
\name{sparse_numeric-class}
\alias{sparse_numeric-class}
\title{Sparse Numeric Vector}
\description{
An S4 class that represents a sparse numeric vector. A sparse numeric vector
is a vector where majority elements are zero.
}
\section{Slots}{

\describe{
\item{\code{value}}{A numeric vector of nonzero values}

\item{\code{pos}}{An integer vector of the positions of nonzero values}

\item{\code{length}}{An integer representing the total length of vector, w/non zeros}
}}

