% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6.R
\name{-,sparse_numeric,sparse_numeric-method}
\alias{-,sparse_numeric,sparse_numeric-method}
\title{subtract one sparse vector from another}
\usage{
\S4method{-}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{e1}{The first sparse vector}

\item{e2}{The second sparse vector, of the same length as the first, that is subtracting from the first.}
}
\value{
Sparse vector of the result after subtraction
}
\description{
This function subtracts one sparse vector, of the same length, by another sparse vector, and then returns the remainder left of the original sparse vector
}
