% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{sparse_sub}
\alias{sparse_sub}
\alias{sparse_sub,sparse_numeric,sparse_numeric-method}
\alias{-,sparse_numeric,sparse_numeric-method}
\title{Subtract Two Sparse Numeric Vectors}
\usage{
sparse_sub(x, y, ...)

\S4method{sparse_sub}{sparse_numeric,sparse_numeric}(x, y, ...)

\S4method{-}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{y}{A sparse_numeric object}

\item{...}{Additional arguments, not used}

\item{e1}{sparse_numeric object}

\item{e2}{sparse_numeric object}
}
\value{
sparse_numeric object containing difference
}
\description{
subtraction of two sparse vector
}
