% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_numeric.R
\name{standardize}
\alias{standardize}
\alias{standardize,sparse_numeric-method}
\title{Standardize Sparse Numeric Vector}
\usage{
standardize(x, ...)

\S4method{standardize}{sparse_numeric}(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{A sparse_numeric object}

\item{...}{Additional arguments (not used)}

\item{na.rm}{logical, should missing values be removed?}
}
\value{
standardized sparse_numeric object
}
\description{
Standardize sparse numeric vector, by subtracting mean and dividing by standard deviation.
}
