% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6_Class.R
\name{standardize}
\alias{standardize}
\title{Standardize a sparse_numeric vector}
\usage{
standardize(x)
}
\arguments{
\item{x}{generic sparse_standardize}
}
\description{
Generic function for standardizing a sparse_numeric vector
}
