% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework6.R
\name{standardize}
\alias{standardize}
\alias{standardize,sparse_numeric-method}
\title{Standardize a sparse_numeric vector}
\usage{
standardize(x, ...)

\S4method{standardize}{sparse_numeric}(x, ...)
}
\arguments{
\item{x}{a \code{sparse_numeric} object.}

\item{...}{ignored.}
}
\value{
A new \code{sparse_numeric} object containing standardized values
}
\description{
Centers and scales a \code{sparse_numeric} vector by subtracting its mean
and dividing by its standard deviation
}
