% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW6.R
\name{standardize}
\alias{standardize}
\alias{standardize,sparse_numeric-method}
\title{standardizes a sparse vector}
\usage{
standardize(first, ...)

\S4method{standardize}{sparse_numeric}(first)
}
\arguments{
\item{first}{The sparse vector}

\item{...}{Not Necessary}
}
\value{
sparse vector
}
\description{
This function standardizes a sparse vector, by taking each element of the vector and subtracting off the vector mean and dividing by the vector standard deviation.
}
